Alternative solution for PYTHON-836
